requirements.txt
Crawl4AI
python-dotenv
pydantic
pyyaml
requests
rich